+++ pytermor/template.py
@@ -8,13 +8,18 @@ from __future__ import annotations
 
 import re
 import typing as t
+from abc import abstractmethod, ABCMeta
+from collections import deque
 from dataclasses import dataclass
+from typing import Union
 
-from .color import resolve_color
-from .common import ExtendedEnum
+from .common import get_qname
+from .color import resolve_color, DEFAULT_COLOR
 from .exception import LogicError
+from .filter import fit
+from .ansi import SequenceCSI, SeqIndex
 from .log import _logger, measure
-from .style import MergeMode, NOOP_STYLE, Style
+from .style import MergeMode, NOOP_STYLE, Style, merge_styles
 from .term import (
     make_clear_display,
     make_clear_display_after_cursor,
@@ -27,37 +32,221 @@ from .term import (
 from .text import Fragment, Text, apply_style_words_selective, apply_style_selective
 
 
-class _TemplateTagOption(str, ExtendedEnum):
-    STYLE_WORDS_SELECTIVE_WSPACE = "|"
-    STYLE_WORDS_SELECTIVE_COMMA = ","
+_T = t.TypeVar("_T")
+
+
+def _tpleng_debug_sep(f=40, n=40):
+    _logger.debug(fit("-" * f, n, align="<"))
+
+
+def _tpleng_debug(l, s):
+    _logger.debug(f"{l:>9s}  {s}")
+
+
+class _StyleSplitter(metaclass=ABCMeta):
+    def __init__(self, tag_style: Style):
+        self._tag_style = tag_style
+
+    @property
+    def tag_style(self) -> Style:
+        return self._tag_style
+
+    @abstractmethod
+    def apply(self, tpl_text: str) -> t.Sequence[Fragment]:
+        raise NotImplementedError
+
+
+class _SplitterWordsSelectiveWSpace(_StyleSplitter):
+    def apply(self, tpl_text: str) -> t.Sequence[Fragment]:
+        return apply_style_words_selective(tpl_text, self._tag_style)
+
+
+class _SplitterWordsSelectiveComma(_StyleSplitter):
+    def apply(self, tpl_text: str) -> t.Sequence[Fragment]:
+        return apply_style_selective(
+            re.compile(r"([^,]+)?([, ]*)"), tpl_text, self._tag_style
+        )
 
 
 @dataclass(frozen=True)
-class _TemplateTagRaw:
-    persist: str | None
+class _TemplateTagGroups:
+    register: str | None
     name: str | None
-    options: str
+    split: str
     action: str
     attrs: str
     pos: str
     clear: str
 
 
+class _TemplateTag:
+    _REGISTER_TO_MERGE_MODE = {
+        "?": MergeMode.FALLBACK,
+        "!": MergeMode.OVERWRITE,
+        "@": MergeMode.REPLACE,
+    }
+    _CLEAR_TO_SEQ = {
+        "<>": make_clear_line(),
+        "<<>>": make_clear_display(),
+        "<": make_clear_line_before_cursor(),
+        ">": make_clear_line_after_cursor(),
+        "<<": make_clear_display_before_cursor(),
+        ">>": make_clear_display_after_cursor(),
+    }
+    _SPLITTERS = {
+        "|": _SplitterWordsSelectiveWSpace,
+        ",": _SplitterWordsSelectiveComma,
+    }
+
+    _ACTION_OPENER = "+"
+    _ACTION_TERMINATOR = "-"
+    _ACTION_RESETTER = "$"
+
+    def __init__(self, raw: str, **groups):
+        self._raw = raw
+        self._groups = _TemplateTagGroups(**groups)
+
+    @property
+    def raw(self) -> str:
+        return self._raw
+
+    @property
+    def groups(self) -> _TemplateTagGroups:
+        return self._groups
+
+    @property
+    def reg_merge_mode(self) -> MergeMode | None:
+        return self._REGISTER_TO_MERGE_MODE.get(self._groups.register, None)
+
+    @property
+    def clear_seq(self) -> SequenceCSI:
+        return self._CLEAR_TO_SEQ.get(self._groups.clear, None)
+
+    @property
+    def is_opener(self) -> bool:
+        return self._groups.action in (self._ACTION_OPENER, "")
+
+    @property
+    def is_terminator(self) -> bool:
+        return self._groups.action == self._ACTION_TERMINATOR
+
+    @property
+    def is_resetter(self) -> bool:
+        return self._groups.action == self._ACTION_RESETTER
+
+    @property
+    def splitter(self) -> type[_StyleSplitter] | None:
+        if not self._groups.split:
+            return None
+        if self._groups.split not in self._SPLITTERS.keys():  # pragma: no cover
+            raise LogicError(
+                f"Invalid template tag style splitter: '{self._groups.split}'"
+            )
+        return self._SPLITTERS.get(self._groups.split)
+
+    def style(self, custom_styles: t.Dict[str, Style]) -> Style | None:
+        if not self._groups.attrs:
+            return NOOP_STYLE
+
+        style_attrs = {}
+        base_style = NOOP_STYLE
+        for style_attr in self._groups.attrs.split(" "):
+            if not style_attr:
+                continue
+            if style_attr in custom_styles.keys():
+                if base_style != NOOP_STYLE:
+                    raise LogicError(f"Max. one custom style per tag: ({style_attr})")
+                base_style = custom_styles[style_attr]
+                continue
+
+            if style_attr.startswith("fg") or style_attr.startswith("bg"):
+                style_attrs.update(
+                    {
+                        k: (v or DEFAULT_COLOR)
+                        for (k, _, v) in (style_attr.partition("="),)
+                    }
+                )
+                continue
+
+            if style_attr in Style.renderable_attributes:
+                style_attrs.update({style_attr: True})
+                continue
+
+            try:
+                color = resolve_color(style_attr)
+                style_attrs.update({"fg": color})
+                continue
+            except LookupError:
+                pass
+
+            raise ValueError(f'Unknown style name or attribute: "{style_attr}"')
+
+        return Style(base_style, **style_attrs)
+
+
+class _LoggingStack(deque[_T], metaclass=ABCMeta):
+    def __init__(self, name: str, *, maxlen: int = None):
+        super().__init__(maxlen=maxlen)
+        self._name = name
+
+    def append(self, st: _T) -> None:
+        _tpleng_debug(f"{self._name} <=", self._repr_item(st))
+        super().append(st)
+
+    def pop(self) -> _T:
+        st = super().pop()
+        _tpleng_debug(f"{self._name} =>", self._repr_item(st))
+        return st
+
+    def clear(self) -> None:
+        _tpleng_debug(f"{self._name} <=", "[]")
+        super().clear()
+
+    def __repr__(self) -> str:
+        return ",".join(map(self._repr_item, self))
+
+    @abstractmethod
+    def _repr_item(self, o: _T) -> str:
+        raise NotImplementedError
+
+
+class _FragmentStack(_LoggingStack[Fragment]):
+    def _repr_item(self, o: Fragment) -> str:
+        return str(o)
+
+
+class _StyleStack(_LoggingStack[Style]):
+    @property
+    def current(self) -> Style:
+        return merge_styles(NOOP_STYLE, overwrites=[*self])
+
+    def _repr_item(self, o: Style) -> str:
+        return o.repr_attrs(verbose=False)
+
+
+class _SplitterStack(_LoggingStack[_StyleSplitter]):
+    def __init__(self, name: str) -> None:
+        super().__init__(name, maxlen=1)
+
+    def _repr_item(self, o: _StyleSplitter) -> str:
+        return get_qname(o)
+
+
 class TemplateEngine:
     _TAG_REGEX = re.compile(
         r"""                  ##### TAG PREFIX #######
-        ((?P<persist>[?!@])   # save tag to local map as fallback/overwrite/explicitly
+        ((?P<register>[?!@])  # save tag to local map as fallback/overwrite/explicitly
          (?P<name>[\w-]+)     # … under specified name  
         )?                    # … or make a one-use tag
         :                     # … and assemble the SGR and insert the bytes
         (?![^\\]\\)           # ignore escaped with single backslash, but not double
         \[(?:                 ####### TAG BODY #######
-            (?P<pos>0)            # reset cursor
+            (?P<pos>\^)           # reset cursor
             |                     # … OR clear current line / the whole screen
-            (?P<clear>(<<|<|<>|<<>>|>|>>))  #
+            (?P<clear>(<<|<|<>|<<>>|>|>>))   #
             |
-            (?P<options>[|,]*)    # … OR /with options/ 
-            (?P<action>[+-]?)     # … open or close style  /empty=open/
+            (?P<split>[|,]?)      # … OR /with splitting/ 
+            (?P<action>[+$-]?)    # … open style/reset styles/close style  (empty=open)
             (?P<attrs>[a-zA-Z][\w =-]*|)   # … add style attributes OR close last  
             
         )\]
@@ -67,118 +256,124 @@ class TemplateEngine:
     _COMMENT_REGEX = re.compile(r"#\[.*?\]")
     _ESCAPE_REGEX = re.compile(r"([^\\])\\\[")
 
-    _PERSIST_TO_MERGE_MODE = {
-        "?": MergeMode.FALLBACK,
-        "!": MergeMode.OVERWRITE,
-        "@": MergeMode.REPLACE,
-    }
-    _CLEAR_TO_SEQ = {
-        "<>": make_clear_line().assemble(),
-        "<<>>": make_clear_display().assemble(),
-        "<": make_clear_line_before_cursor().assemble(),
-        ">": make_clear_line_after_cursor().assemble(),
-        "<<": make_clear_display_before_cursor().assemble(),
-        ">>": make_clear_display_after_cursor().assemble(),
-    }
-
     def __init__(self, custom_styles: t.Dict[str, Style] = None):
         self._user_styles: t.Dict[str, Style] = custom_styles or {}
+
+    def reset(self):
+        self._user_styles.clear()
 
     @measure(
         template_enter_fn=lambda _, tpl: f"Substituting the template ({len(tpl)})",
         template_exit="Substituted in %s",
     )
     def substitute(self, tpl: str) -> Text:
-        result = Text()
+        _logger.debug(f"Substituting the template (L={len(tpl)})")
+        tpl_parts = self._split(tpl)
+
+        _logger.debug(f"Split to {len(tpl_parts)} parts")
+        return self._assemble(tpl_parts)
+
+    def _split(self, tpl: str) -> deque[Union[str, _TemplateTag]]:
         tpl_cursor = 0
-        style_buffer = NOOP_STYLE
-        st_opts = []
-        _logger.debug(f"Substituting the template ({len(tpl)})")
-
-        #@TODO rework into stacked-type iter
         tpl_nocom = self._COMMENT_REGEX.sub("", tpl)
-        for tag_match in self._TAG_REGEX.finditer(tpl_nocom):
-            span = tag_match.span()
-            tpl_part = self._ESCAPE_REGEX.sub(r"\1[", tpl_nocom[tpl_cursor : span[0]])
-            if len(tpl_part) > 0 or style_buffer != NOOP_STYLE:
-                selective_fn = None
-                if _TemplateTagOption.STYLE_WORDS_SELECTIVE_WSPACE in st_opts:
-                    selective_fn = apply_style_words_selective(tpl_part, style_buffer)
-                elif _TemplateTagOption.STYLE_WORDS_SELECTIVE_COMMA in st_opts:
-                    selective_fn = apply_style_selective(re.compile(r'([^,]+)?([, ]*)'), tpl_part, style_buffer)
-                if selective_fn:
-                    for part in selective_fn:
-                        result += part
+        tpl_parts = deque[Union[str, _TemplateTag]]()
+
+        for idx, tag_match in enumerate(self._TAG_REGEX.finditer(tpl_nocom)):
+            tpl_span = tag_match.span()
+            text_before = self._ESCAPE_REGEX.sub(
+                r"\1[", tpl_nocom[tpl_cursor : tpl_span[0]]
+            )
+            tpl_cursor = tpl_span[1]
+            if text_before:
+                tpl_parts.append(text_before)
+            tpl_parts.append(_TemplateTag(tag_match.group(), **tag_match.groupdict()))
+
+        if text_last := tpl_nocom[tpl_cursor:]:
+            tpl_parts.append(text_last)
+
+        return tpl_parts
+
+    def _assemble(self, tpl_parts: deque[Union[str, _TemplateTag]]) -> Text:
+        result = _FragmentStack("result")
+        st_stack = _StyleStack("stack")
+        spl_stack = _SplitterStack("spl st")
+
+        from_stack = st_stack.pop
+        to_stack = st_stack.append
+
+        for idx, tpl_part in enumerate(tpl_parts):
+            _stacks_str = f"[" + "] [".join(repr(s) for s in [spl_stack, st_stack]) + "]"
+            _tpleng_debug_sep()
+            _tpleng_debug("STACKS", _stacks_str)
+
+            if isinstance(tpl_part, str):
+                _tpleng_debug(f"PART {idx}", f"{tpl_part!r}")
+
+                if spl_stack:
+                    _tpleng_debug("TEXT MODE", "split_stack")
+                    for segm in (splitter := spl_stack.pop()).apply(tpl_part):
+                        result.append(segm)
                     # add open style for engine to properly handle the :[-closing] tag:
-                    tpl_part = ""
-                result += Fragment(tpl_part, style_buffer, close_this=False)
+                    result.append(Fragment("", splitter.tag_style, close_this=False))
+                    continue
 
-            tpl_cursor = span[1]
-            tagr = _TemplateTagRaw(**tag_match.groupdict())
-            if tagr.action == "-" and not tagr.attrs:
-                result += Fragment("", style_buffer, close_prev=True)
-            elif tagr.pos:
-                result += Fragment(make_reset_cursor().assemble(), None)
-            elif tagr.clear:
-                result += Fragment(self._CLEAR_TO_SEQ.get(tagr.clear, ""), None)
-
-            style_buffer = NOOP_STYLE
-            st_opts = []
-
-            input_style = self._tag_to_style(tagr, self._user_styles)
-            if merge_mode := self._PERSIST_TO_MERGE_MODE.get(tagr.persist):
-                self._user_styles[tagr.name] = self._user_styles.get(
-                    tagr.name, NOOP_STYLE.clone()
-                ).merge(merge_mode, input_style)
+                _tpleng_debug("TEXT MODE", "stack")
+                result.append(Fragment(tpl_part, st_stack.current))
                 continue
 
-            if tagr.action == "-":
-                result += Fragment("", input_style, close_prev=True)
-            else:
-                style_buffer = input_style
-                st_opts = [
-                    {v: k for k, v in _TemplateTagOption.dict().items()}.get(opt)
-                    for opt in (tagr.options or [])
-                ]
-            continue
+            tag = t.cast(_TemplateTag, tpl_part)
+            tag_style = tag.style(self._user_styles)
+            _tpleng_debug(f"PART {idx}", f"{tag.raw}")
 
-        result += tpl_nocom[tpl_cursor:]
-        return result
+            if splitter := tag.splitter:
+                _tpleng_debug("TAG TYPE", "split")
+                spl_stack.append(splitter(tag_style))
 
-    def _tag_to_style(
-        self, raw: _TemplateTagRaw, custom_styles: t.Dict[str, Style]
-    ) -> Style | None:
-        style_attrs = {}
-        base_style = NOOP_STYLE
-        if raw.attrs:
-            for style_attr in raw.attrs.split(" "):
-                if not style_attr:
+            elif tag.reg_merge_mode:
+                _tpleng_debug("TAG TYPE", "register")
+                existing_st = self._user_styles.get(tag.groups.name, NOOP_STYLE.clone())
+                updated_st = existing_st.merge(tag.reg_merge_mode, tag_style)
+                _tpleng_debug("userst <=", updated_st)
+                self._user_styles[tag.groups.name] = updated_st
+
+            elif tag.groups.pos:
+                _tpleng_debug("TAG TYPE", "reset_cursor")
+                result.append(Fragment(make_reset_cursor().assemble()))
+
+            elif tag.groups.clear:
+                _tpleng_debug("TAG TYPE", "clear_seq")
+                result.append(Fragment(tag.clear_seq.assemble()))
+
+            elif tag.is_resetter:
+                _tpleng_debug("TAG TYPE", "reset")
+                st_stack.clear()
+                result.append(Fragment(SeqIndex.RESET.assemble()))
+
+            elif tag.is_terminator:
+                if not st_stack:
+                    _tpleng_debug(
+                        "TAG TYPE", "terminator | IGNORED because style stack is empty"
+                    )
                     continue
-                if style_attr in custom_styles.keys():
-                    if base_style != NOOP_STYLE:
+                if not tag.groups.attrs:
+                    _tpleng_debug("TAG TYPE", "terminator_prev")
+                    from_stack()
+                else:
+                    _tpleng_debug("TAG TYPE", "terminator_exact")
+                    if (st_prev := from_stack()) != tag_style:
                         raise LogicError(
-                            f"Only one custom style per tag is allowed: ({style_attr})"
+                            f"Terminator '{tag.raw}' doesn't match corresponding opener: {st_prev}"
                         )
-                    base_style = custom_styles[style_attr]
-                    continue
 
-                if style_attr.startswith("fg=") or style_attr.startswith("bg="):
-                    style_attrs.update({k: v for k, v in (style_attr.split("="),)})
-                    continue
+            elif tag.is_opener:
+                _tpleng_debug("TAG TYPE", "opener")
+                to_stack(tag_style)
 
-                if style_attr in Style.renderable_attributes:
-                    style_attrs.update({style_attr: True})
-                    continue
+            else:  # pragma: no cover
+                raise LogicError(f"Invalid tag: {tag.raw}")
 
-                try:
-                    if color := resolve_color(style_attr):
-                        style_attrs.update({"fg": color})
-                        continue
-                except LookupError:
-                    pass
-
-                raise ValueError(f'Unknown style name or attribute: "{style_attr}"')
-        return Style(base_style, **style_attrs)
+        _tpleng_debug_sep()
+        return Text(*result)
 
 
 _template_engine = TemplateEngine()